migration of old tests to new response validation directory
